TGP-3173: update packages and split controller for goodsRecord.CountryOfOriginController
