+++ app/navigation/GoodsRecordNavigator.scala
@@ -37,7 +37,7 @@ class GoodsRecordNavigator @Inject() extends Navigator {
     case p: ProductReferenceUpdatePage    =>
       _ => controllers.goodsRecord.routes.CyaUpdateRecordController.onPageLoadproductReference(p.recordId)
     case GoodsDescriptionPage             =>
-      _ => controllers.goodsRecord.routes.CountryOfOriginController.onPageLoadCreate(NormalMode)
+      _ => controllers.goodsRecord.countryOfOrigin.routes.CreateCountryOfOriginController.onPageLoad(NormalMode)
     case p: GoodsDescriptionUpdatePage    =>
       _ => controllers.goodsRecord.routes.CyaUpdateRecordController.onPageLoadGoodsDescription(p.recordId)
     case CountryOfOriginPage              =>
@@ -90,7 +90,7 @@ class GoodsRecordNavigator @Inject() extends Navigator {
       .get(HasCountryOfOriginChangePage(recordId))
       .map {
         case false => controllers.goodsRecord.routes.SingleRecordController.onPageLoad(recordId)
-        case true  => controllers.goodsRecord.routes.CountryOfOriginController.onPageLoadUpdate(NormalMode, recordId)
+        case true  => controllers.goodsRecord.countryOfOrigin.routes.UpdateCountryOfOriginController.onPageLoad(NormalMode, recordId)
       }
       .getOrElse(controllers.problem.routes.JourneyRecoveryController.onPageLoad(Some(continueUrl)))
   }
